Adicionado busca nos endpoins por todos os campos da entidade, e alimentada a documentacao no swaager

+++ src/main/java/com/example/bluebank/model/Lancamentos.java
@@ -1,5 +1,7 @@
 package com.example.bluebank.model;
 
+import java.math.BigDecimal;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -16,11 +18,11 @@ public class Lancamentos {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY.AUTO)
-	private int id;
+	private Integer id;
 	@Column(name = "c_idcliente",columnDefinition = "integer")
-	private int idcliente;
+	private Integer idcliente;
 	@Column(name = "c_idconta",columnDefinition = "integer")
-	private int idconta;
+	private Integer idconta;
 	@Column(name = "c_dataevento",columnDefinition = "varchar(10)")
 	private String data_evento;
 	@Column(name = "c_tipo",columnDefinition = "varchar(1)")
@@ -28,6 +30,6 @@ public class Lancamentos {
 	@Column(name = "c_motivo",columnDefinition = "varchar(100)")
 	private String motivo;
 	@Column(name = "c_valor",columnDefinition = "numeric(13,2)")
-	private double valor;
+	private BigDecimal valor;
 
 }
